requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
python-dotenv>=1.0
requests>=2.31

chromadb>=0.5
sentence-transformers>=2.6
rank-bm25>=0.2
numpy>=1.26
scikit-learn>=1.4

prometheus-client>=0.20
opentelemetry-api>=1.24
opentelemetry-sdk>=1.24
opentelemetry-exporter-otlp-proto-http>=1.24
opentelemetry-instrumentation-fastapi>=0.45b0
opentelemetry-instrumentation-requests>=0.45b0

streamlit>=1.36
